build: relocate binutils symlinks
Since the SDK no longer has a single target architecture, it's not
possible to add the helper symlinks for binutils to $PATH so that
RPM's `find-debuginfo.sh` will automatically use the right programs
to extract debug symbols from binaries built for the target.

Instead, prepare directories that can be added to $PATH based on the
desired target architecture. This will need to happen in downstream
code, prior to building the RPMs.

Signed-off-by: Ben Cressey <bcressey@amazon.com>