build: convert ARG to ENV where possible
ARG is intended for build-time image customization, while ENV is for
runtime customization. Both directives affect the environment of any
subsequent layer that depends on them.

Since the SDK is ultimately copied into a "scratch" base layer, the
ENV directives do not actually affect the environment when the SDK
is used for Bottlerocket builds.

Switch to ENV for any value that is static and not expected to change
based on build parameters. That way, ARG always refers exclusively to
inputs coming from the `make` invocation, which makes it easier to
understand where the build has conditional behavior.

Signed-off-by: Ben Cressey <bcressey@amazon.com>