build: compress SDK with zstd
Rather than implicitly using BuildKit underneath `docker build`,
switch to explicitly using it via `docker buildx build` with a custom
builder.

The default builder loads builds into Docker after they finish, which
causes certain options - like zstd compression - to be ignored when
pushing to a registry.

`docker buildx build` doesn't really distinguish between "build" and
"push" steps; a "push" is just a build where the output is sent to a
registry rather than written to a tar archive or loaded into Docker.
This breaks one of the main assumptions of the `publish-sdk` script,
which expects the build to be done already.

Rather than wiring up the build arguments as additional arguments to
`publish-sdk`, replace it with `docker buildx imagetools create` as
the tool for creating and replacing remote manifests.

Signed-off-by: Ben Cressey <bcressey@amazon.com>